+++ apps/laboratory/tests/smart-account.spec.ts
@@ -13,31 +13,59 @@ testModalSmartAccount.beforeEach(async ({ modalValidator }) => {
   await modalValidator.expectConnected()
 })
 
-testModalSmartAccount('it should sign with eoa', async ({ modalPage, modalValidator }) => {
-  await modalPage.sign()
-  await modalPage.approveSign()
-  await modalValidator.expectAcceptedSign()
+testModalSmartAccount('it should use a Smart Account', async ({ modalPage, modalValidator }) => {
+  const walletModalPage = modalPage as ModalWalletPage
+  const walletModalValidator = modalValidator as ModalWalletValidator
+
+  await modalValidator.expectConnected()
+  await walletModalPage.openAccount()
+  await walletModalValidator.expectActivateSmartAccountPromoVisible(false)
+
+  await walletModalPage.openProfileView()
+  await walletModalPage.openSettings()
+  await walletModalValidator.expectChangePreferredAccountToShow(EOA)
+  await walletModalPage.closeModal()
+})
+
+testModalSmartAccount('it should sign with 6492', async ({ modalPage, modalValidator }) => {
+  const walletModalPage = modalPage as ModalWalletPage
+  const walletModalValidator = modalValidator as ModalWalletValidator
+
+  await walletModalPage.sign()
+  await walletModalPage.approveSign()
+  await walletModalValidator.expectAcceptedSign()
+
+  const signature = await walletModalPage.getSignature()
+  const address = await walletModalPage.getAddress()
+  const chainId = await walletModalPage.getChainId()
+
+  await walletModalValidator.expectValidSignature(signature, address, chainId)
 })
 
 testModalSmartAccount(
-  'it should switch to its smart account and sign',
+  'it should switch to its eoa and sign',
   async ({ modalPage, modalValidator }) => {
     const walletModalPage = modalPage as ModalWalletPage
     const walletModalValidator = modalValidator as ModalWalletValidator
 
     await walletModalPage.openAccount()
-    await walletModalValidator.expectActivateSmartAccountPromoVisible(true)
-
+    await walletModalPage.openProfileView()
     await walletModalPage.openSettings()
-    await walletModalValidator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
+
     await walletModalPage.togglePreferredAccountType()
-    await walletModalValidator.expectChangePreferredAccountToShow(EOA)
+    await walletModalValidator.expectChangePreferredAccountToShow(SMART_ACCOUNT)
 
     await walletModalPage.closeModal()
 
     await walletModalPage.sign()
     await walletModalPage.approveSign()
     await walletModalValidator.expectAcceptedSign()
+
+    const signature = await walletModalPage.getSignature()
+    const address = await walletModalPage.getAddress()
+    const chainId = await walletModalPage.getChainId()
+
+    await walletModalValidator.expectValidSignature(signature, address, chainId)
   }
 )
 
@@ -50,10 +78,9 @@ testModalSmartAccount(
     const originalAddress = await walletModalPage.getAddress()
 
     await walletModalPage.openAccount()
+    await walletModalPage.openProfileView()
     await walletModalPage.openSettings()
 
-    await walletModalPage.togglePreferredAccountType()
-    await walletModalValidator.expectChangePreferredAccountToShow(EOA)
     await walletModalPage.switchNetwork('Avalanche')
     await modalValidator.expectSwitchedNetwork('Avalanche')
     await walletModalValidator.expectTogglePreferredTypeVisible(false)
@@ -63,28 +90,6 @@ testModalSmartAccount(
     await walletModalValidator.expectActivateSmartAccountPromoVisible(false)
     await walletModalPage.closeModal()
 
-    await walletModalValidator.expectAddress(originalAddress)
-  }
-)
-
-testModalSmartAccount(
-  'it should properly sign with a 6492 signature',
-  async ({ modalPage, modalValidator }) => {
-    const walletModalPage = modalPage as ModalWalletPage
-    const walletModalValidator = modalValidator as ModalWalletValidator
-
-    await walletModalPage.openAccount()
-    await walletModalPage.openSettings()
-    await walletModalPage.togglePreferredAccountType()
-    await walletModalValidator.expectChangePreferredAccountToShow(EOA)
-    await walletModalPage.closeModal()
-
-    await walletModalPage.sign()
-    await walletModalPage.approveSign()
-    await walletModalValidator.expectAcceptedSign()
-    const signature = await walletModalPage.getSignature()
-    const address = await walletModalPage.getAddress()
-    const chainId = await walletModalPage.getChainId()
-    await walletModalValidator.expectValidSignature(signature, address, chainId)
+    await walletModalValidator.expectChangedAddressAfterSwitchingAccountType(originalAddress)
   }
 )