+++ packages/core/src/controllers/OptionsController.ts
@@ -18,9 +18,11 @@ export interface OptionsControllerState {
   privacyPolicyUrl?: string
   isSiweEnabled?: boolean
   isSiwsEnabled?: boolean
+  isUniversalProvider?: boolean
   enableAnalytics?: boolean
   metadata?: Metadata
   enableOnramp?: boolean
+  hasMultipleAddresses?: boolean
   disableAppend?: boolean
   enableEIP6963?: boolean
 }
@@ -85,6 +87,10 @@ export const OptionsController = {
     state.isSiweEnabled = isSiweEnabled
   },
 
+  setIsUniversalProvider(isUniversalProvider: OptionsControllerState['isUniversalProvider']) {
+    state.isUniversalProvider = isUniversalProvider
+  },
+
   setIsSiwsEnabled(isSiwsEnabled: OptionsControllerState['isSiwsEnabled']) {
     state.isSiwsEnabled = isSiwsEnabled
   },
@@ -111,5 +117,9 @@ export const OptionsController = {
 
   setEIP6963Enabled(enableEIP6963: OptionsControllerState['enableEIP6963']) {
     state.enableEIP6963 = enableEIP6963
+  },
+
+  setHasMultipleAddresses(hasMultipleAddresses: OptionsControllerState['hasMultipleAddresses']) {
+    state.hasMultipleAddresses = hasMultipleAddresses
   }
 }