+++ packages/scaffold-ui/src/views/w3m-connecting-wc-view/index.ts
@@ -75,7 +75,12 @@ export class W3mConnectingWcView extends LitElement {
 
         await ConnectionController.connectWalletConnect()
         this.finalizeConnection()
-        if (OptionsController.state.isSiweEnabled) {
+        if (
+          StorageUtil.getConnectedConnector() === 'AUTH' &&
+          OptionsController.state.hasMultipleAddresses
+        ) {
+          RouterController.push('SelectAddresses')
+        } else if (OptionsController.state.isSiweEnabled) {
           const { SIWEController } = await import('@web3modal/siwe')
           if (SIWEController.state.status === 'success') {
             ModalController.close()
@@ -133,7 +138,7 @@ export class W3mConnectingWcView extends LitElement {
     const { mobile_link, desktop_link, webapp_link, injected, rdns } = this.wallet
     const injectedIds = injected?.map(({ injected_id }) => injected_id).filter(Boolean) as string[]
     const browserIds = rdns ? [rdns] : injectedIds ?? []
-    const isBrowser = browserIds.length
+    const isBrowser = OptionsController.state.isUniversalProvider ? false : browserIds.length
     const isMobileWc = mobile_link
     const isWebWc = webapp_link
     const isBrowserInstalled = ConnectionController.checkInstalled(browserIds)